requirements.txt
transformers
opencv-python
Pillow
numpy
pandas
onnxruntime
huggingface_hub